Renaming NewConstants to Database
